requirements.txt
click>=8.0.0
rich>=13.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
